OS detection Patch & File/IO Introduction
Introductions:
🟢 File class introduction with low-level operations for high speed file managment.
🟢 Introduced AMPLE.h a configuration file.

Changes:
🔵 Fixed OS and Architecture detection through MACROS in AMPLE.h.
🔵 Now AMPLE can be compiled in all OSs but with introduce warning message when not compiled in Windows MSVC or Linux GCC/G++.
🔵 Now all warnings will be treated as errors.

Known issues:s
🔴 Dynamic libraries can not be loaded in Mac OS.

Future improvements:
🟡 Library installation.
🟡 Will introduce Load Program and Program Matrices into Memory functions.

+++ Libraries/Memory/include/Memory.h
@@ -18,8 +18,8 @@
 class Memory
 {
 public:
-    EXPORT static bool InitializeMemory(uint8_t ssize, uint64_t nsize);
-    EXPORT static bool ResizeMemory(uint64_t nsize);
+    EXPORT static bool InitializeMemory(uint8_t ssize, size_t nsize);
+    EXPORT static bool ResizeMemory(size_t nsize);
     EXPORT static void FreeMemory();
     EXPORT static void* Get(uint64_t pos);
 };
